5.9 Renaming
Changed:
Renamed some things to be more organized.
For example:
PlayerInventoryController is main script, with refereces to PlayerInventory_Ammo and PlayerInventory_Weapon which are still this inventory, but seperated. Something like sub controllers.

Base bobbing is now seperated into weapon and camera.
Weapon bobbing is calculated when player isnt aiming and camera when player is aiming.