6.4 New way to pickup
Changed:
Now instead of using raycast to pickup things. Now the closest pickupable to the Pickuper will light us its outline and be set as the currect pickup. This is the one that will be put in inventory on interaction input.

CanvasGroup toggle is now seperated into its own class created on other scripts awake.
Now only in this script toggle logic has to be changed.

Added:
Interaction HUD(disabled for now).

OnMoveUpdate for mainTransformer.

Fixed:
Trigger discipline not disabling on drop and temporaryUnEquip.