7.7.2 Combining stackable items
Added:
Combining stackable items.
If two item slots hold the same  stackable items and the one that is targeted has space left for more items.
The dragged item will give as much items to the targeted slot and empty itself if all items were given to it.
Only impelemented for I-I (Item to items) drag and drop.