8.3 RightHandFollow
Changed:
LeftHandIk can now be toggled to inherid the rightHands position and rotation or not.

LeftHand will only move when lerping, which removes the need for positioner.

Added:
Put together a simple system for throwing items with animation. It works but doesn't look the best.
I might need to reconsider using baked animations and ik combination, but pay more attention to these animations.