9.2.2 Equip animations
Added:
Weapon equip animation system.
Instead of just lerping rightHandIk to place, it is now set instantly. It is not visible to the player, since baked layer is enabled, allowing to play animations and use events to trigger method on anim end.

Changed:
Adjusted equping in combatsController Equip sub-controller to correctly work with new animations.