Fixed gravity-slope-weapon bug
Fixed:
Weapon behaving weirdly on a slope.
Being on a slope caused inAir weapon animator to rotate weapon to fall rotation(because slopes increase gravity, so that player can walk on them).
Now slopeAngle still affects players gravity, but gravityStrength of inAir animator is 0 when grounded. This allows player to walk on slopes without any weird weapon behaviours.

Note:
Tested climbing with weapon, and climbing which lead to conclusion that climbing might have to be refreshed/upgraded.

I have only tested and fixed today because of leetcode and neatcode practice.