10.0 Dynamic vault
Changed:
ClimbController shoots multiple rays in down direction each further away from the player to check if obsticle can be climbed correctly, and if it can be vaulted over. So there is no need for vault tag anymore, script checks it with raycasts.

Added:
VeryHigh climb, but because of lack of animations, High climb is used for now to reach higher obstacles. Better animations will be added in the future.